Update PE_smallest_multiple.cpp
fixed data types.

+++ Lesson_4/PE_smallest_multiple.cpp
@@ -1,8 +1,8 @@
 #include <iostream>
 
-uint64_t greatest_common_divisor(long a, long b) { //Euclidean  Algorithm: taken from google
+uint64_t greatest_common_divisor(uint64_t a, uint64_t b) { //Euclidean  Algorithm: taken from google
     while (b != 0) {
-        int old_b = b;
+        uint64_t old_b = b;
         b = a % b;
         a = old_b;
         //std::cout << "old b: " << old_b << std::endl;
@@ -13,14 +13,14 @@ uint64_t greatest_common_divisor(long a, long b) { //Euclidean  Algorithm: taken
     return a;
 }
 
-long least_common_multiple(long a, long b) {
+uint64_t least_common_multiple(uint64_t a, uint64_t b) {
     return a * (b / greatest_common_divisor(a, b));
 }
 
-long smallest_multiple(long num) {
-    long result   = num;
-    long next_dig = num - 1;
-    for (long next_dig = result - 1; next_dig > 1; --next_dig) {
+uint64_t smallest_multiple(uint64_t num) {
+    uint64_t result   = num;
+    uint64_t next_dig = num - 1;
+    for (uint64_t next_dig = result - 1; next_dig > 1; --next_dig) {
         result = least_common_multiple(result, next_dig);
     }
     std::cout << result << std::endl;
